server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Sample server with autotests, database, JWT, versions and logger routing"
authors = ["I.B. <mail@mail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = {extras = ["brotli"], version = "^0.23.3"}
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.8"
alembic = "^1.10.2"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
asyncpg = "^0.27.0"
simplejson = "^3.18.4"
pytest = "^7.2.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pytest-asyncio = "^0.21.0"
pika = "^1.3.1"
loguru = "^0.6.0"
stackprinter = "^0.2.10"
aio-pika = "^9.0.5"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

